apres rebase + correction geoJson
