+++ routes/graph.js
@@ -1,220 +1,219 @@
 const debug = require('debug')('graph');
-const debugPatch = require('debug')('patch');
+// const debugPatch = require('debug')('patch');
 const router = require('express').Router();
 const fs = require('fs');
 const path = require('path');
-const { matchedData, query, body } = require('express-validator');
+const { matchedData, query } = require('express-validator');
 const jimp = require('jimp');
-const PImage = require('pureimage');
 
-const GJV = require('geojson-validation');
+// const GJV = require('geojson-validation');
 const validateParams = require('../paramValidation/validateParams');
-const validator = require('../paramValidation/validator');
+// const validator = require('../paramValidation/validator');
 const createErrMsg = require('../paramValidation/createErrMsg');
 
-const geoJsonAPatcher = [
-  body('geoJSON')
-    .exists().withMessage(createErrMsg.missingBody)
-    .custom(GJV.isGeoJSONObject)
-    .withMessage(createErrMsg.invalidBody('objet GeoJSON'))
-    .custom(GJV.isFeatureCollection)
-    .withMessage(createErrMsg.invalidBody('featureCollection')),
-  body('geoJSON.type')
-    .exists().withMessage(createErrMsg.missingParameter('type'))
-    .isIn(['FeatureCollection'])
-    .withMessage(createErrMsg.invalidParameter('type')),
-  body('geoJSON.crs')
-    .exists().withMessage(createErrMsg.missingParameter('crs'))
-    .custom(validator.isCrs)
-    .withMessage(createErrMsg.invalidParameter('crs')),
-  body('geoJSON.features.*.geometry')
-    .custom(GJV.isPolygon).withMessage(createErrMsg.InvalidEntite('geometry', 'polygon')),
-  body('geoJSON.features.*.properties.color')
-    .exists().withMessage(createErrMsg.missingParameter('properties.color'))
-    .custom(validator.isColor)
-    .withMessage(createErrMsg.invalidParameter('properties.color')),
-  body('geoJSON.features.*.properties.cliche')
-    .exists().withMessage(createErrMsg.missingParameter('properties.cliche'))
-    .matches(/^[a-zA-Z0-9-_]+$/i)
-    .withMessage(createErrMsg.invalidParameter('properties.cliche')),
-];
+// const geoJsonAPatcher = [
+//   body('geoJSON')
+//     .exists().withMessage(createErrMsg.missingBody)
+//     .custom(GJV.isGeoJSONObject)
+//     .withMessage(createErrMsg.invalidBody('objet GeoJSON'))
+//     .custom(GJV.isFeatureCollection)
+//     .withMessage(createErrMsg.invalidBody('featureCollection')),
+//   body('geoJSON.type')
+//     .exists().withMessage(createErrMsg.missingParameter('type'))
+//     .isIn(['FeatureCollection'])
+//     .withMessage(createErrMsg.invalidParameter('type')),
+//   body('geoJSON.crs')
+//     .exists().withMessage(createErrMsg.missingParameter('crs'))
+//     .custom(validator.isCrs)
+//     .withMessage(createErrMsg.invalidParameter('crs')),
+//   body('geoJSON.features.*.geometry')
+//     .custom(GJV.isPolygon).withMessage(createErrMsg.InvalidEntite('geometry', 'polygon')),
+//   body('geoJSON.features.*.properties.color')
+//     .exists().withMessage(createErrMsg.missingParameter('properties.color'))
+//     .custom(validator.isColor)
+//     .withMessage(createErrMsg.invalidParameter('properties.color')),
+//   body('geoJSON.features.*.properties.cliche')
+//     .exists().withMessage(createErrMsg.missingParameter('properties.cliche'))
+//     .matches(/^[a-zA-Z0-9-_]+$/i)
+//     .withMessage(createErrMsg.invalidParameter('properties.cliche')),
+// ];
 
-// Encapsulation des informations du requestBody dans une nouvelle clé 'keyName' ("body" par defaut)
-function encapBody(req, res, next) {
-  let keyName = 'body';
-  if (this.keyName) { keyName = this.keyName; }
-  if (JSON.stringify(req.body) !== '{}') {
-    const requestBodyKeys = Object.keys(req.body);
-    req.body[keyName] = JSON.parse(JSON.stringify(req.body));
-    for (let i = 0; i < requestBodyKeys.length; i += 1) {
-      delete req.body[requestBodyKeys[i]];
-    }
-  }
-  next();
-}
+// // Encapsulation des informations du requestBody dans une nouvelle clé 'keyName' ("body" par defaut)
+// function encapBody(req, res, next) {
+//   let keyName = 'body';
+//   if (this.keyName) { keyName = this.keyName; }
+//   if (JSON.stringify(req.body) !== '{}') {
+//     const requestBodyKeys = Object.keys(req.body);
+//     req.body[keyName] = JSON.parse(JSON.stringify(req.body));
+//     for (let i = 0; i < requestBodyKeys.length; i += 1) {
+//       delete req.body[requestBodyKeys[i]];
+//     }
+//   }
+//   next();
+// }
 
-router.post('/graph/patch', encapBody.bind({ keyName: 'geoJSON' }), [
-  ...geoJsonAPatcher,
-], validateParams, (req, res) => {
-  const params = matchedData(req);
-  const X0 = 0;
-  const Y0 = 12000000;
-  const R = 0.05;
-  const geoJson = params.geoJSON;
-  const promises = [];
+// router.post('/graph/patch', encapBody.bind({ keyName: 'geoJSON' }), [
+//   ...geoJsonAPatcher,
+// ], validateParams, (req, res) => {
+//   const params = matchedData(req);
+//   const X0 = 0;
+//   const Y0 = 12000000;
+//   const R = 0.05;
+//   const geoJson = params.geoJSON;
+//   const promises = [];
 
-  debug('GeoJson:');
-  debug(geoJson);
-  debug('Features:');
-  debug(geoJson.features);
-  debug(geoJson.features[0].geometry.coordinates);
-  debug(geoJson.features[0].properties.color);
+//   debug('GeoJson:');
+//   debug(geoJson);
+//   debug('Features:');
+//   debug(geoJson.features);
+//   debug(geoJson.features[0].geometry.coordinates);
+//   debug(geoJson.features[0].properties.color);
 
-  // BBox du polygone a patcher
-  const BBox = {};
-  try {
-    geoJson.features.forEach((feature) => {
-      feature.geometry.coordinates[0].forEach((point) => {
-        if ('xmin' in BBox) {
-          BBox.xmin = Math.min(BBox.xmin, point[0]);
-          BBox.xmax = Math.max(BBox.xmax, point[0]);
-          BBox.ymin = Math.min(BBox.ymin, point[1]);
-          BBox.ymax = Math.max(BBox.ymax, point[1]);
-        } else {
-          [BBox.xmin, BBox.ymin] = point;
-          [BBox.xmax, BBox.ymax] = point;
-        }
-      });
-    });
-    debug('BBox:');
-    debug(BBox);
+//   // BBox du polygone a patcher
+//   const BBox = {};
+//   try {
+//     geoJson.features.forEach((feature) => {
+//       feature.geometry.coordinates[0].forEach((point) => {
+//         if ('xmin' in BBox) {
+//           BBox.xmin = Math.min(BBox.xmin, point[0]);
+//           BBox.xmax = Math.max(BBox.xmax, point[0]);
+//           BBox.ymin = Math.min(BBox.ymin, point[1]);
+//           BBox.ymax = Math.max(BBox.ymax, point[1]);
+//         } else {
+//           [BBox.xmin, BBox.ymin] = point;
+//           [BBox.xmax, BBox.ymax] = point;
+//         }
+//       });
+//     });
+//     debug('BBox:');
+//     debug(BBox);
 
-    // List of all tiles
-    const tiles = [];
-    let resolution = R;
-    const errors = [];
+//     // List of all tiles
+//     const tiles = [];
+//     let resolution = R;
+//     const errors = [];
 
-    for (let z = 21; z >= 10; z -= 1) {
-      const x0 = Math.floor((BBox.xmin - X0) / (resolution * 256));
-      const x1 = Math.ceil((BBox.xmax - X0) / (resolution * 256));
-      const y0 = Math.floor((Y0 - BBox.ymax) / (resolution * 256));
-      const y1 = Math.ceil((Y0 - BBox.ymin) / (resolution * 256));
-      for (let y = y0; y < y1; y += 1) {
-        for (let x = x0; x < x1; x += 1) {
-          tiles.push({
-            x, y, z, resolution,
-          });
+//     for (let z = 21; z >= 10; z -= 1) {
+//       const x0 = Math.floor((BBox.xmin - X0) / (resolution * 256));
+//       const x1 = Math.ceil((BBox.xmax - X0) / (resolution * 256));
+//       const y0 = Math.floor((Y0 - BBox.ymax) / (resolution * 256));
+//       const y1 = Math.ceil((Y0 - BBox.ymin) / (resolution * 256));
+//       for (let y = y0; y < y1; y += 1) {
+//         for (let x = x0; x < x1; x += 1) {
+//           tiles.push({
+//             x, y, z, resolution,
+//           });
 
-          const pathGraph = path.join(global.dir_cache, `${z}`, `${y}`, `${x}`, 'graph.png');
-          const pathOrtho = path.join(global.dir_cache, `${z}`, `${y}`, `${x}`, 'ortho.png');
-          const pathOpi = path.join(global.dir_cache, `${z}`, `${y}`, `${x}`, `${geoJson.features[0].properties.cliche}.png`);
-          if (!fs.existsSync(pathGraph) || !fs.existsSync(pathOrtho) || !fs.existsSync(pathOpi)) {
-            errors.push(`${path.join(global.dir_cache, `${z}`, `${y}`, `${x}`)}`);
-          }
-        }
-      }
-      resolution *= 2;
-    }
+//           const pathGraph = path.join(global.dir_cache, `${z}`, `${y}`, `${x}`, 'graph.png');
+//           const pathOrtho = path.join(global.dir_cache, `${z}`, `${y}`, `${x}`, 'ortho.png');
+//           const pathOpi = path.join(global.dir_cache, `${z}`, `${y}`, `${x}`, `${geoJson.features[0].properties.cliche}.png`);
+//           if (!fs.existsSync(pathGraph) || !fs.existsSync(pathOrtho) || !fs.existsSync(pathOpi)) {
+//             errors.push(`${path.join(global.dir_cache, `${z}`, `${y}`, `${x}`)}`);
+//           }
+//         }
+//       }
+//       resolution *= 2;
+//     }
 
-    if (errors.length) {
-      debugPatch('ERROR');
-      debugPatch(errors);
-      const err = new Error();
-      err.code = 404;
-      err.msg = {
-        status: 'Fichier(s) absent(s)',
-        errors: [{
-          param: 'graph.png, ortho.png et/ou {nomCliche}.png',
-          localisation: errors,
-        }],
-      };
-      throw err;
-    }
+//     if (errors.length) {
+//       debugPatch('ERROR');
+//       debugPatch(errors);
+//       const err = new Error();
+//       err.code = 404;
+//       err.msg = {
+//         status: 'Fichier(s) absent(s)',
+//         errors: [{
+//           param: 'graph.png, ortho.png et/ou {nomCliche}.png',
+//           localisation: errors,
+//         }],
+//       };
+//       throw err;
+//     }
 
-    debugPatch(tiles);
-    // Patch these tiles
-    tiles.forEach((tile) => {
-    // Patch du graph
-      debugPatch(tile);
-      const urlGraph = path.join(global.dir_cache, `${tile.z}`, `${tile.y}`, `${tile.x}`, 'graph.png');
-      const urlOrtho = path.join(global.dir_cache, `${tile.z}`, `${tile.y}`, `${tile.x}`, 'ortho.png');
-      const urlOpi = path.join(global.dir_cache, `${tile.z}`, `${tile.y}`, `${tile.x}`, `${geoJson.features[0].properties.cliche}.png`);
+//     debugPatch(tiles);
+//     // Patch these tiles
+//     tiles.forEach((tile) => {
+//     // Patch du graph
+//       debugPatch(tile);
+//       const urlGraph = path.join(global.dir_cache, `${tile.z}`, `${tile.y}`, `${tile.x}`, 'graph.png');
+//       const urlOrtho = path.join(global.dir_cache, `${tile.z}`, `${tile.y}`, `${tile.x}`, 'ortho.png');
+//       const urlOpi = path.join(global.dir_cache, `${tile.z}`, `${tile.y}`, `${tile.x}`, `${geoJson.features[0].properties.cliche}.png`);
 
-      const mask = PImage.make(256, 256);
-      const ctx = mask.getContext('2d');
-      geoJson.features.forEach((feature) => {
-        debugPatch(feature.properties.color);
-        ctx.fillStyle = '#FFFFFF';
-        ctx.beginPath();
-        let first = true;
-        /* eslint-disable no-restricted-syntax */
-        for (const point of feature.geometry.coordinates[0]) {
-          const i = Math.round((point[0] - X0 - tile.x * 256 * tile.resolution) / tile.resolution);
-          const j = Math.round((Y0 - point[1] - tile.y * 256 * tile.resolution) / tile.resolution);
-          // debugPatch(i, j);
-          if (first) {
-            first = false;
-            ctx.moveTo(i, j);
-          } else {
-            ctx.lineTo(i, j);
-          }
-        }
-        ctx.closePath();
-        ctx.fill();
-      });
+//       const mask = PImage.make(256, 256);
+//       const ctx = mask.getContext('2d');
+//       geoJson.features.forEach((feature) => {
+//         debugPatch(feature.properties.color);
+//         ctx.fillStyle = '#FFFFFF';
+//         ctx.beginPath();
+//         let first = true;
+//         /* eslint-disable no-restricted-syntax */
+//         for (const point of feature.geometry.coordinates[0]) {
+//           const i = Math.round((point[0] - X0 - tile.x * 256 * tile.resolution) / tile.resolution);
+//           const j = Math.round((Y0 - point[1] - tile.y * 256 * tile.resolution) / tile.resolution);
+//           // debugPatch(i, j);
+//           if (first) {
+//             first = false;
+//             ctx.moveTo(i, j);
+//           } else {
+//             ctx.lineTo(i, j);
+//           }
+//         }
+//         ctx.closePath();
+//         ctx.fill();
+//       });
 
-      // // On patch le graph
-      /* eslint-disable no-param-reassign */
-      promises.push(jimp.read(urlGraph).then((graph) => {
-        for (let idx = 0; idx < 256 * 256 * 4; idx += 4) {
-          if (mask.data[idx + 3]) {
-            [graph.bitmap.data[idx],
-              graph.bitmap.data[idx + 1],
-              graph.bitmap.data[idx + 2]] = geoJson.features[0].properties.color;
-          }
-        }
-        return graph.writeAsync(urlGraph);
-      }).then(() => {
-        debugPatch(`${urlGraph}: done`);
-      }));
-      // // On patch l ortho
-      /* eslint-disable no-param-reassign */
-      const promiseOrthoOpi = [jimp.read(urlOrtho), jimp.read(urlOpi)];
-      promises.push(Promise.all(promiseOrthoOpi).then((images) => {
-        const ortho = images[0];
-        const opi = images[1];
-        // debug(ortho, opi);
-        for (let idx = 0; idx < 256 * 256 * 4; idx += 4) {
-          if (mask.data[idx + 3]) {
-            ortho.bitmap.data[idx] = opi.bitmap.data[idx];
-            ortho.bitmap.data[idx + 1] = opi.bitmap.data[idx + 1];
-            ortho.bitmap.data[idx + 2] = opi.bitmap.data[idx + 2];
-          }
-        }
-        return ortho.writeAsync(urlOrtho);
-      }).then(() => {
-        debugPatch('done');
-      }));
-    });
+//       // // On patch le graph
+//       /* eslint-disable no-param-reassign */
+//       promises.push(jimp.read(urlGraph).then((graph) => {
+//         for (let idx = 0; idx < 256 * 256 * 4; idx += 4) {
+//           if (mask.data[idx + 3]) {
+//             [graph.bitmap.data[idx],
+//               graph.bitmap.data[idx + 1],
+//               graph.bitmap.data[idx + 2]] = geoJson.features[0].properties.color;
+//           }
+//         }
+//         return graph.writeAsync(urlGraph);
+//       }).then(() => {
+//         debugPatch(`${urlGraph}: done`);
+//       }));
+//       // // On patch l ortho
+//       /* eslint-disable no-param-reassign */
+//       const promiseOrthoOpi = [jimp.read(urlOrtho), jimp.read(urlOpi)];
+//       promises.push(Promise.all(promiseOrthoOpi).then((images) => {
+//         const ortho = images[0];
+//         const opi = images[1];
+//         // debug(ortho, opi);
+//         for (let idx = 0; idx < 256 * 256 * 4; idx += 4) {
+//           if (mask.data[idx + 3]) {
+//             ortho.bitmap.data[idx] = opi.bitmap.data[idx];
+//             ortho.bitmap.data[idx + 1] = opi.bitmap.data[idx + 1];
+//             ortho.bitmap.data[idx + 2] = opi.bitmap.data[idx + 2];
+//           }
+//         }
+//         return ortho.writeAsync(urlOrtho);
+//       }).then(() => {
+//         debugPatch('done');
+//       }));
+//     });
 
-    debug('execution des promises');
-    Promise.all(promises)
-      .then(() => {
-        debug("tout c'est bien passé");
-        res.status(200).send(JSON.stringify(tiles));
-      })
-      .catch((erreur) => {
-        debug('erreur : ', erreur);
-        // todo: il faut tout annuler
-        const err = new Error();
-        err.code = 500;
-        err.msg = err;
-        throw err;
-      });
-  } catch (err) {
-    res.status(err.code).send(err.msg);
-  }
-});
+//     debug('execution des promises');
+//     Promise.all(promises)
+//       .then(() => {
+//         debug("tout c'est bien passé");
+//         res.status(200).send(JSON.stringify(tiles));
+//       })
+//       .catch((erreur) => {
+//         debug('erreur : ', erreur);
+//         // todo: il faut tout annuler
+//         const err = new Error();
+//         err.code = 500;
+//         err.msg = err;
+//         throw err;
+//       });
+//   } catch (err) {
+//     res.status(err.code).send(err.msg);
+//   }
+// });
 
 router.get('/graph', [
   query('x')
@@ -232,13 +231,20 @@ router.get('/graph', [
   const { y } = params;
 
   debug(x, y);
-  const X = 0;
-  const Y = 12000000;
-  const R = 178.571428571429 * 0.00028;
+
+  // const X = 0;
+  // const Y = 12000000;
+  // const R = 178.571428571429 * 0.00028;
+  let fullRes;
+  req.app.tileSet.forEach((level) => {
+    if ((!fullRes) || (fullRes.resolution > level.resolution)) {
+      fullRes = level;
+    }
+  });
 
   // il faut trouver la tuile
-  const Px = (x - X) / R;
-  const Py = (Y - y) / R;
+  const Px = (x - fullRes.x0) / fullRes.resolution;
+  const Py = (fullRes.y0 - y) / fullRes.resolution;
   const Tx = Math.floor(Px / 256);
   const Ty = Math.floor(Py / 256);
   const I = Math.floor(Px - Tx * 256);
